examples/image: use tokio's current_thread runtime
This is part of an experiment to find unexplained memory consumption.
My suspicsion is that there's a leak when using
tokio_core::reactor::Core directly as we did here. Since the the
error_chain error's don't implement the Send trait, the future chain
cannot be executed on tokio::runtime::Runtime, but must run on the
tokio::runtime::current_thread::Runtime instead.